refactor: single with_predicate for delete ops
