+++ rust/src/operations/mod.rs
@@ -24,9 +24,9 @@ pub mod vacuum;
 #[cfg(feature = "datafusion")]
 use self::{delete::DeleteBuilder, load::LoadBuilder, update::UpdateBuilder, write::WriteBuilder};
 #[cfg(feature = "datafusion")]
-use arrow::record_batch::RecordBatch;
+pub use ::datafusion::physical_plan::common::collect as collect_sendable_stream;
 #[cfg(feature = "datafusion")]
-pub use datafusion::physical_plan::common::collect as collect_sendable_stream;
+use arrow::record_batch::RecordBatch;
 #[cfg(all(feature = "arrow", feature = "parquet"))]
 use optimize::OptimizeBuilder;
 
@@ -166,5 +166,35 @@ impl From<DeltaOps> for DeltaTable {
 impl AsRef<DeltaTable> for DeltaOps {
     fn as_ref(&self) -> &DeltaTable {
         &self.0
+    }
+}
+
+#[cfg(feature = "datafusion")]
+mod datafusion {
+    use datafusion_expr::Expr;
+
+    /// Used to represent user input of either a Datafusion expression or string expression
+    pub enum Expression {
+        /// Datafusion Expression
+        DataFusion(Expr),
+        /// String Expression
+        String(String),
+    }
+
+    impl From<Expr> for Expression {
+        fn from(val: Expr) -> Self {
+            Expression::DataFusion(val)
+        }
+    }
+
+    impl From<&str> for Expression {
+        fn from(val: &str) -> Self {
+            Expression::String(val.to_string())
+        }
+    }
+    impl From<String> for Expression {
+        fn from(val: String) -> Self {
+            Expression::String(val)
+        }
     }
 }
